Split-off 'PrinterCore' from 'Printer.hh'
